+++ src/parser/Ast.ts
@@ -8,7 +8,7 @@ import {CellError} from '../Cell'
 import {AddressWithSheet} from './Address'
 import {CellAddress} from './CellAddress'
 import {ColumnAddress} from './ColumnAddress'
-import {IExtendedToken} from './FormulaParser'
+import {ExtendedToken} from './FormulaParser'
 import {RowAddress} from './RowAddress'
 
 export type Ast =
@@ -141,7 +141,7 @@ export interface StringAst extends AstWithWhitespace {
   value: string,
 }
 
-export const buildStringAst = (token: IExtendedToken): StringAst => ({
+export const buildStringAst = (token: ExtendedToken): StringAst => ({
   type: AstNodeType.STRING,
   value: token.image.slice(1, -1),
   leadingWhitespace: token.leadingWhitespace?.image,