+++ src/parser/ColumnAddress.ts
@@ -35,6 +35,10 @@ export class ColumnAddress implements AddressWithColumn {
     return new ColumnAddress(ReferenceType.RELATIVE, column, sheet)
   }
 
+  public static compareByAbsoluteAddress(baseAddress: SimpleCellAddress): (colA: ColumnAddress, colB: ColumnAddress) => number {
+    return (colA: ColumnAddress, colB: ColumnAddress) =>  colA.toSimpleColumnAddress(baseAddress).col - colB.toSimpleColumnAddress(baseAddress).col
+  }
+
   public isColumnAbsolute(): boolean {
     return (this.type === ReferenceType.ABSOLUTE)
   }
@@ -75,7 +79,7 @@ export class ColumnAddress implements AddressWithColumn {
     return new ColumnAddress(this.type, col, this.sheet)
   }
 
-  public withAbsoluteSheet(sheet: number): ColumnAddress {
+  public withSheet(sheet: number | undefined): ColumnAddress {
     return new ColumnAddress(this.type, this.col, sheet)
   }