+++ src/parser/RowAddress.ts
@@ -30,6 +30,10 @@ export class RowAddress implements AddressWithRow {
     return new RowAddress(ReferenceType.RELATIVE, row, sheet)
   }
 
+  public static compareByAbsoluteAddress(baseAddress: SimpleCellAddress): (rowA: RowAddress, rowB: RowAddress) => number {
+    return (rowA: RowAddress, rowB: RowAddress) =>  rowA.toSimpleRowAddress(baseAddress).row - rowB.toSimpleRowAddress(baseAddress).row
+  }
+
   public isRowAbsolute(): boolean {
     return (this.type === ReferenceType.ABSOLUTE)
   }
@@ -70,7 +74,7 @@ export class RowAddress implements AddressWithRow {
     return new RowAddress(this.type, row, this.sheet)
   }
 
-  public withAbsoluteSheet(sheet: number): RowAddress {
+  public withSheet(sheet: number | undefined): RowAddress {
     return new RowAddress(this.type, this.row, sheet)
   }