+++ test/i18n.spec.ts
@@ -2,7 +2,7 @@ import {HyperFormula, LanguageAlreadyRegisteredError, LanguageNotRegisteredError
 import {ProtectedFunctionTranslationError} from '../src/errors'
 import {RawTranslationPackage, TranslationPackage} from '../src/i18n'
 import * as languages from '../src/i18n/languages'
-import {plPL} from '../src/i18n/languages'
+import {enGB, plPL} from '../src/i18n/languages'
 import {FunctionRegistry} from '../src/interpreter/FunctionRegistry'
 import {CellAddress} from '../src/parser'
 import {adr, extractReference} from './testUtils'
@@ -144,5 +144,13 @@ describe('i18n', () => {
     expect(() =>
       HyperFormula.getLanguage('foo')
     ).toThrow(new LanguageNotRegisteredError())
+  })
+
+  describe('registerLanguage', () => {
+    it('should throw error when language code is not a string', () => {
+      expect(() => {
+        HyperFormula.registerLanguage(42 as any, enGB)
+      }).toThrowError('Expected value of type: string for config parameter: languageCode')
+    })
   })
 })