+++ test/parser/white-spaces.spec.ts
@@ -3,9 +3,9 @@ import {SheetMapping} from '../../src/DependencyGraph'
 import {buildTranslationPackage} from '../../src/i18n'
 import {enGB} from '../../src/i18n/languages'
 import {buildLexerConfig, FormulaLexer} from '../../src/parser'
-import {CellReference, EqualsOp, ProcedureName, RangeSeparator, RParen, WhiteSpace} from '../../src/parser/LexerConfig'
-import {bindWhitespacesToTokens} from '../../src/parser/ParserWithCaching'
+import {CellReference, EqualsOp, ProcedureName, RangeSeparator, RParen} from '../../src/parser/LexerConfig'
 import {expectArrayWithSameContent} from '../testUtils'
+import {buildEmptyParserWithCaching} from './common'
 
 describe('tokenizeFormula', () => {
   const config = new Config()
@@ -20,8 +20,7 @@ describe('tokenizeFormula', () => {
 
   it('should forget about trailing whitespaces', () => {
     const tokens = lexer.tokenizeFormula('=SUM(A1:A2)   ').tokens
-    expect(tokens.length).toBe(6)
-    expect(tokens.map(token => token.tokenType).find(tokenType => tokenType === WhiteSpace)).toBe(undefined)
+    expect(tokens.map(token => token.tokenType.name)).not.toContain('WhiteSpace')
   })
 
   it('should skip whitespace inside range', () => {
@@ -65,29 +64,37 @@ describe('tokenizeFormula', () => {
 
   it('should treat space as whitespace', () => {
     const tokens = lexer.tokenizeFormula('= 1').tokens
-    expect(tokens[1].tokenType).toEqual(WhiteSpace)
+    expect(tokens[1].tokenType.name).toEqual('WhiteSpace')
   })
 
   it('should treat tabulator (U+0009) as whitespace', () => {
     const tokens = lexer.tokenizeFormula('=\t1').tokens
-    expect(tokens[1].tokenType).toEqual(WhiteSpace)
+    expect(tokens[1].tokenType.name).toEqual('WhiteSpace')
   })
 
   it('should treat carriage return (U+000D) as whitespace', () => {
     const tokens = lexer.tokenizeFormula('=\r1').tokens
-    expect(tokens[1].tokenType).toEqual(WhiteSpace)
+    expect(tokens[1].tokenType.name).toEqual('WhiteSpace')
   })
 
   it('should treat line feed (U+000A) as whitespace ', () => {
     const tokens = lexer.tokenizeFormula('=\n1').tokens
-    expect(tokens[1].tokenType).toEqual(WhiteSpace)
+    expect(tokens[1].tokenType.name).toEqual('WhiteSpace')
   })
 
   it('should treat multiple whitespaces as one token', () => {
     const tokens = lexer.tokenizeFormula('=\n\t\r 1').tokens
     expect(tokens.length).toEqual(3)
-    expect(tokens[1].tokenType).toEqual(WhiteSpace)
+    expect(tokens[1].tokenType.name).toEqual('WhiteSpace')
     expect(tokens[1].image).toEqual('\n\t\r ')
+  })
+
+  it('when set ignoreWhiteSpace = \'any\', should treat non-breaking space as a whitespace', () => {
+    const config = new Config({ ignoreWhiteSpace: 'any' })
+    const lexer = new FormulaLexer(buildLexerConfig(config))
+
+    const tokens = lexer.tokenizeFormula('=\u00A042').tokens
+    expect(tokens[1].tokenType.name).toEqual('WhiteSpace')
   })
 })
 
@@ -95,11 +102,11 @@ describe('processWhitespaces', () => {
   const config = new Config()
   const sheetMapping = new SheetMapping(buildTranslationPackage(enGB))
   sheetMapping.addSheet('Sheet1')
-  const lexer = new FormulaLexer(buildLexerConfig(config))
+  const parser = buildEmptyParserWithCaching(config, sheetMapping)
 
   it('should do nothing when no whitespaces', () => {
-    const tokens = lexer.tokenizeFormula('=SUM(A1:A2)').tokens
-    const processed = bindWhitespacesToTokens(tokens)
+    const tokens = parser.tokenizeFormula('=SUM(A1:A2)').tokens
+    const processed = parser.bindWhitespacesToTokens(tokens)
     expect(processed.length).toBe(6)
     expect(processed.map(processed => processed.leadingWhitespace).every(processed => processed === undefined)).toBe(true)
     expectArrayWithSameContent(
@@ -109,8 +116,8 @@ describe('processWhitespaces', () => {
   })
 
   it('should add leading whitespace to token', () => {
-    const tokens = lexer.tokenizeFormula('= SUM(A1:A2)').tokens
-    const processed = bindWhitespacesToTokens(tokens)
+    const tokens = parser.tokenizeFormula('= SUM(A1:A2)').tokens
+    const processed = parser.bindWhitespacesToTokens(tokens)
     expect(processed.length).toBe(6)
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     expect(processed[1].leadingWhitespace!.image).toBe(' ')
@@ -121,8 +128,8 @@ describe('processWhitespaces', () => {
   })
 
   it('should work for multiple whitespaces', () => {
-    const tokens = lexer.tokenizeFormula('=    SUM(A1:A2)').tokens
-    const processed = bindWhitespacesToTokens(tokens)
+    const tokens = parser.tokenizeFormula('=    SUM(A1:A2)').tokens
+    const processed = parser.bindWhitespacesToTokens(tokens)
     expect(processed.length).toBe(6)
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     expect(processed[1].leadingWhitespace!.image).toBe('    ')
@@ -133,8 +140,8 @@ describe('processWhitespaces', () => {
   })
 
   it('should work for whitespace at the beginning', () => {
-    const tokens = lexer.tokenizeFormula(' =SUM(A1:A2)').tokens
-    const processed = bindWhitespacesToTokens(tokens)
+    const tokens = parser.tokenizeFormula(' =SUM(A1:A2)').tokens
+    const processed = parser.bindWhitespacesToTokens(tokens)
     expect(processed.length).toBe(6)
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     expect(processed[0].leadingWhitespace!.image).toBe(' ')
@@ -145,8 +152,8 @@ describe('processWhitespaces', () => {
   })
 
   it('should not include whitespaces directly on the list', () => {
-    const tokens = lexer.tokenizeFormula('=   SUM(   A1:A2)   ').tokens
-    const processed = bindWhitespacesToTokens(tokens)
+    const tokens = parser.tokenizeFormula('=   SUM(   A1:A2)   ').tokens
+    const processed = parser.bindWhitespacesToTokens(tokens)
     expect(processed.length).toBe(6)
     expectArrayWithSameContent(
       [EqualsOp, ProcedureName, CellReference, RangeSeparator, CellReference, RParen],